minor performance improvement

+++ include/utils.h
@@ -40,10 +40,10 @@
 	const char *ltrim(const char *src);
 
 	unsigned long ahtoi(const char *data);
-	int atoi_s(const char *str, long long length);
+	int atoi_s(const char *str, short length);
 	char *base64_encode(const char *data, const unsigned long data_length);
 
-	long char_at(const char *data, const char ch, unsigned long length);
+	int char_at(const char *data, const char ch, int length);
 
 	unsigned long long get_timestamp();
